requirements.txt
Flask>=2.2.0
requests>=2.28.0
Flask-Cors>=3.0.10
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
python-dotenv>=1.0.0
